requirements.txt
antlr4-python3-runtime==4.8
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.27
hydra-core==1.1.2
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.3
omegaconf==2.1.2
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.1.0
pip-autoremove==0.10.0
platformdirs==2.5.2
promise==2.3
protobuf==3.20.1
psutil==5.9.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.16
